fairseq-py/examples/nllb/mining/global_mining/requirements.txt
# you'll need fairseq too

# useful for the NLLBModules
hydra-core >= 1.1.0
submitit
tqdm

# for the global mining logic
numpy
faiss-gpu
torch

# if we want to use the Hugging Face transformers
transformers
